pyproject.toml
[project]
name = "intelligent-conversational-agent"
version = "0.1.0"
description = "Conversational AI system that simulates an intelligent agent for a fictional business"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.12",
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "streamlit>=1.55.0",
    "yoyo-migrations>=9.0.0",
]
